Add AsyncRPCInterface in AsyncCache, improve unit tests
